edit mission fail

+++ Mars Exploration Simulator/Rovers/Rover.h
@@ -30,8 +30,8 @@ public:
 	void reset();
 
 	//setters
-	virtual void assignMission(int idMission,int duration, int location,int curDay);
-	virtual void assignCheckup(int curDay);
+	virtual void assignMission(int idMission,int duration, int location, unsigned long long curDay);
+	virtual void assignCheckup(unsigned long long curDay);
 	virtual void decreaseSpeedToHalf();
 	virtual void increaseSpeedToDouble();
 	virtual void assignMaintainance(int);
